reorganize map.js, split into separate files
